Extract resize_img to image_utils.py, breaking circular import
Move `resize_img` and `_resize_pil_image` from `mast3r_utils.py` into a
new leaf module `image_utils.py` (depends only on PIL, numpy, dust3r).
This breaks the circular import chain (frame -> mast3r_utils -> frame)
so `mast3r_utils.py` can now import `Frame` directly and use proper
type annotations instead of `object`.

- New: `mast3r_slam/image_utils.py`
- `mast3r_utils.py`: import Frame, replace all `object` -> `Frame`,
  remove unused PIL/ImgNorm imports
- `frame.py`, `dataloader.py`: import resize_img from image_utils
