+++ packages/mast3r-slam/mast3r_slam/frame.py
@@ -8,7 +8,7 @@ from numpy import ndarray
 from torch import Tensor
 
 from mast3r_slam.config import config
-from mast3r_slam.mast3r_utils import resize_img
+from mast3r_slam.image_utils import resize_img
 
 
 class Mode(Enum):
